third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-containment-show-all-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests Containment view of detailed heap snapshots. The "Show All" button must show all nodes. Test object distances calculation.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll

Running: testDistances

Profiler was disabled.
